type conversion warnings: Use 'DataAccessPosition_t' to limit size of data allocation to 2 G and use 'IntervalType' for offsets.
Buffer::number_of_channels is 'int'.

+++ lib/backtrace/tasktimer.cpp
@@ -266,7 +266,7 @@ bool TaskTimer::printIndentation() {
             stringstream ss;
 
             int width = 1;
-            int N = thread_info_map.size ();
+            size_t N = thread_info_map.size ();
             int number = t.threadNumber;
 
             while ((N/=10) > 1)
