+++ lib/tfrheightmap/heightmap/update/tfrblockupdater.cpp
@@ -78,7 +78,7 @@ TfrBlockUpdater::Job::Job(Tfr::pChunk chunk, float nf, float largest_fs)
     normalization_factor *= 100.f;
 
     Tfr::ChunkElement *cp = chunk->transform_data->getCpuMemory ();
-    int n = chunk->transform_data->numberOfElements ();
+    int n = (int)chunk->transform_data->numberOfElements ();
     // Compute the norm of the complex elements in the chunk prior to resampling and interpolating
     float* fp = computeNorm(cp, n, normalization_factor);
 
@@ -100,7 +100,7 @@ TfrBlockUpdater::Job::Job(Tfr::pChunk chunk, float nf, float largest_fs)
     case Tfr::Chunk::Order_row_major: // i.e Cwt
         org_width = chunk->nSamples ();
         org_height = chunk->nScales ();
-        data_width = int_div_ceil (chunk->n_valid_samples, stepx);
+        data_width = (int)int_div_ceil (chunk->n_valid_samples, stepx);
         data_height = org_height;
         offs_x = chunk->first_valid_sample; // != 0
         break;
